# Conflicts:
#	lib/screens/lawyer_screen/screens/controllers/lawyer_appointment_controllers.dart

+++ lib/screens/client_screen/services/model/lawyer_list_model.dart
@@ -50,7 +50,7 @@ class LawyerListData {
   List<String>? languageSpoken;
   String? experience;
   String? image;
-  double? avgRating;
+  int? avgRating;
 
   LawyerListData(
       {this.id,
